Second round of comments
Don't know why I decided to do the comments in two different rounds, but here are the rest of the comments. Made sure that (almost) everything has a description to make sure it is completely understood how each thing works. (Note: Some of these comments may of been a little excessive, but understand it was done with good intentions)